NewThoughtForm component
