--- code/src/components/ThoughtForm.js
@@ -1,1 +1,0 @@
-import React from 'react'
